requirement.txt
flask-bootstrap
flask_sqlalchemy
mysql-python
flask-mysql
Flask-Migrate
Flask-Script
Flask-MySQLdb